fix(compiler): share prototype memory id parsing
